requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
google-generativeai==0.3.2